Added search function
Added the search function, and changed the insert function to use it. Since I am currently preventing duplicate values from being added to the tree, a search must be performed every time the user inserts a value to make sure that value is not already in the list. It is therefore important to have the search function return information that will be useful to the insert function so that unnecessary tree traversals are avoided.