Removing node w/ two children
Added support for the case where the node to be deleted has to child nodes. To do this I first created a function called findSubTreeMin, which takes a node for an argument and finds the node of minimum value in the subtree beginning with that node. This function would return the tree's minimum value if passed the root node. 

When the node to be deleted has two children, the algorithm first runs findSubTreeMin on that nodes right child. Then the node to be deleted has it's value changed to the value of the node returned by findSubTreeMin. Finally, the algorithm deletes the node returned by findSubTreeMin so that there is no duplicate.