added main function
Added the main function to handle user input. User input is given via pop-up window prompts, and feedback is given through pop-up alerts. Implementing this also involved changing instances of console.log(...) to alert(...), as well as modifying the print algorithms a so that the each node of the tree isn't printed in a separate pop-up