split main form up into separate forms
I was hoping that just one big form would be easier to implement since it did not require as many style changes, but it also meant that accessing elements from the form was significantly for complicated. Additionally, splitting the forms up like this makes it easier for me to add/remove form elements if I want to.