Added return statements to insert function
Now when the user performs an insert in to the tree, the insert function returns the string that was previously displayed in a pop-up window. Adding the return statements will allow me to print this string under the "Last Action:" <h3> element. 

I left the alert(...) pop-up in place for the case where the user tries to add a duplicate value, since it makes sense for a page to display a pop-up if you try to perform an invalid action. Additionally, this allows me to more easily have the "Last action:" field display the users most recent valid action.