Fixed insert function
I had forgotten to add code to make sure that only numeric values could be inserted in to the tree. Now, if the user tries to insert a non-numeric value they will be displayed a pop-up. Trying to insert a non-numeric value will not change the content under the "Last Action:" heading.