added support to display search+remove on webpage
Much like for the insert function, now when the user performs a search or remove the field under "Last action:" is updated if the operation was valid. For a non-valid operation the field under "Last Action:" remains unchanged. 

For now I am defining an invalid operation for these functions as a search/remove on an empty tree, an attempted removal of a value not in the tree, or if the user tries to enter a non-numeric value.

Also fixed a bug that previously existed when the user tried to search on an empty tree.