update print functions to reduce pop-ups
Basically just needed to change instances of "/r" to "<br>" and change the alert(treeString) to return treeString. If the tree is empty, I am choosing to leave the "tree is empty" as an alert as opposed to changing it to a returned value. I am doing this to remain consistent with the search and remove functions, which also trigger an alert pop-up if attempted on an empty tree.