Added support to display printed tree on webpage
If the tree is non-empty, the field under "Last Action" is updated to display the printed tree. Same approach as with previous functions, but needed to update the inner HTML instead of the inner text in order to get proper line breaks. 

I am defining trying to print an empty tree as an invalid operation, so attempting to do so will not update the field under "Last Action". I did this to remain consistent with the search and remove functions. Additionally, clicking the print button without selecting a printing order is an invalid operation.