Added canvas element
I plan on implementing a new print method which actually draws the tree in its current state on to the webpage. To do this, I created a canvas element, on which the tree will be drawn. This element is currently set to not display anything. The commented out script portion draws an example tree.