created method to turn tree into array
The most obvious approach to balancing the tree starts with creating an array of the tree's contents. I have implemented the getTreeArray() method to create this array. This method uses an inorder traversal, so the returned array is sorted numerically.

The balanceTree() method currently just tests the getTreeArray method. The getTreeInfo() method is not currently being used, but will be if I figure out that the balanceTree() method requires more information about the tree than just its contents in array form