Basic balance works
Can balance tree w/ no problems if tree contains easy number of elements (i.e. 3, 7, 15). Currently displays warning message if tree cannot be perfectly balanced.

The balance method is iterative. It inserts the median value of the array of values, splits this array in to a right and left half, than performs this same operation on these new arrays. Iterations end when the array passed to the function contains only one value.