added remove all button
In the same form as the normal remove option, I have added a button labeled "remove all". When clicked, the tree's contents are entirely cleared. This button is especially useful for debugging the balance operation, since it means that I no longer need to refresh the page each time I wish to test it