fixed unnecessary alerts when balancing
Added a boolean flag balancing, which is raised when balancing starts and lowered when balancing ends. When raised the balancing flag tells the insert function not to trigger an alert pop-up if passed an invalid input, which happens when performing the balancing operation on a tree that cannot be perfectly balanced.