fixed multi-insert last action
Fixed the last action field for multi-insert to only display last valid action. Before, if the user attempted to insert a duplicate value as a part of a multi-insert, the last action field would display "undefined" for the duplicate value along with an alert pop up. Now only the alert pop-up is triggered and there is no "undefined" displayed.